Refac : enum 타입 entity 내부로 정리
