--- src/main/java/project/closet/domain/users/user/Gender.java
@@ -1,5 +1,0 @@
-package project.closet.domain.users.user;
-
-public enum Gender {
-  MALE, FEMALE, OTHER
-}
