--- src/main/java/project/closet/domain/users/user/UserRole.java
@@ -1,5 +1,0 @@
-package project.closet.domain.users.user;
-
-public enum UserRole {
-  USER, ADMIN
-}
