+++ src/main/java/project/closet/domain/users/user/dto/UserDto.java
@@ -7,7 +7,7 @@ import java.util.UUID;
 import project.closet.common.util.TimeConverter;
 import project.closet.domain.users.auth.OAuthProvider;
 import project.closet.domain.users.user.User;
-import project.closet.domain.users.user.UserRole;
+import project.closet.domain.users.user.User.UserRole;
 
 public record UserDto(
     UUID id,
